requirements.txt
streamlit>=1.28.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
scikit-learn>=1.3.0
joblib>=1.3.0
scipy>=1.10.0
scikit-image>=0.20.0
pydantic>=2.0.0
